Added a new toggle flip-flop that takes a single bool value and toggles the state if that boolean is true. Renamed `Boolean_Latch` to `Boolean_FlipFlopSR`.
